Why save final cause and error
Source code explains why this saves the cause and error of the final context
cancelation rather than the first:

When a context.Context is canceled, it stores the first error and cause
because that is the error and cause that caused it to be
canceled. However, the AllOf derived context is not canceled until all
of its source contexts have been canceled, or in other words, its final
source context is canceled. Therefore, rather than remembering the
first error and cause, this remembers the error and cause of the final
source context.