thousand context text
Added an AllOf test that spawns a number of goroutines, currently set to 1000,
each of which create a context and Add it to the AllOf instance. They each do
some miniscule amount of work by incrementing a shared counter, then
terminate.

The primary test goroutine waits until the AllOf instance's context is
canceled, which only happens after each of the spawned goroutines have
completed their work and canceled their contexts. After the AllOf context is
canceled, the primary test goroutine ensures that the shared counter has been
incremented the specified number of times.