AllOf has Cancel(cause error) method to free resources of instance.
Cancel is used when the derived context is no longer required, and used to
terminate all goroutines spawned by this AllOf instance that track the
cancelation of their respective context.Context instances, and then cancels
the derived context. This method does not return until all spawned
goroutines have been terminated.

NOTE: This method does not in any way cancel or otherwise affect the
context.Context instances passed to this instance's Add method. It only
cancels the derived context and stops the goroutines watching the added
contexts.